move network and network_extension in their own files
