+++ lib/core/network/api/services/remote_config_service.dart
@@ -2,20 +2,7 @@ import 'dart:convert';
 
 import 'package:firebase_remote_config/firebase_remote_config.dart';
 import 'package:flutter/services.dart';
-
-enum Network {
-  telos,
-  telosTestnet,
-  eos,
-  eosTestnet;
-
-  static Network fromString(String label) {
-    return values.firstWhere(
-      (v) => v.name == label,
-      orElse: () => Network.telos,
-    );
-  }
-}
+import 'package:hypha_wallet/core/network/models/network.dart';
 
 /// Encapsulates everything to do with remote configuration
 class RemoteConfigService {
