+++ lib/ui/profile/interactor/profile_data.dart
@@ -1,4 +1,6 @@
+// ignore: unused_import
 import 'package:hypha_wallet/core/network/api/services/remote_config_service.dart';
+import 'package:hypha_wallet/core/network/models/network.dart';
 import 'package:hypha_wallet/core/network/models/user_profile_data.dart';
 
 class ProfileData {
