+++ test/parse_esr_link_test.dart
@@ -3,6 +3,7 @@ import 'package:hypha_wallet/core/crypto/dart_esr/dart_esr.dart';
 import 'package:hypha_wallet/core/crypto/eosdart/src/client.dart';
 import 'package:hypha_wallet/core/network/api/eos_service.dart';
 import 'package:hypha_wallet/core/network/api/services/remote_config_service.dart';
+import 'package:hypha_wallet/core/network/models/network.dart';
 
 import 'mocks/mock_secure_storage_service.dart';
